Lab 04/CatTigerClassifier/requirements.txt
tensorflow==2.20.0
keras==3.11.3
numpy==2.3.3
matplotlib==3.10.6
scikit-learn==1.7.2
pillow==11.3.0
seaborn==0.13.2
requests==2.32.5

